refactor: useAuth from context
