+++ frontend/src/components/RequireAuth.jsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet, useLocation } from 'react-router-dom'
-import useAuth from '@/hooks/useAuth'
+import { useAuth } from '@/contexts/auth'
 import useUserSessions from '@/hooks/useUserSessions'
 import Loading from './Loading'
