+++ frontend/src/components/UserConfig/StudentForm.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useForm } from 'react-hook-form'
 import { MultiSelect } from 'react-multi-select-component'
-import useAuth from '@/hooks/useAuth'
+import { useAuth } from '@/contexts/auth'
 import useCategories from '@/hooks/useCategories'
 import { updateUserData } from '@/services/users'
 import { countryOptions, genderOptions } from './commonOptions'
