+++ frontend/src/hooks/useAxiosPrivate.js
@@ -1,5 +1,5 @@
 import { useEffect } from 'react'
-import useAuth from './useAuth'
+import { useAuth } from '@/contexts/auth'
 import { axiosPrivate } from '@/utils/axios'
 
 function useAxiosPrivate() {
